Move 'next' and 'prev' helpers to ResourceCollection
